+++ app/(protected)/_layout.tsx
@@ -8,7 +8,7 @@ import { Header } from "@/components/navigation/header/Header";
 import { SideMenu } from "@/components/navigation/side-menu/SideMenu";
 import { useSessionContext } from "@/feature/auth/contexts/SessionContext";
 import { useAuthRedirect } from "@/feature/auth/hooks/useAuthRedirect";
-import { LinkInputModalProvider } from "@/feature/links/contexts/LinkInputModalContext";
+import { LinkInputModalProvider } from "@/feature/links/application/contexts/LinkInputModalContext";
 import { LinkInputView } from "@/feature/links/views/LinkInputView";
 import { ProfileEditModalProvider } from "@/feature/user/contexts/ProfileEditModalContext";
 import { UserProvider } from "@/feature/user/contexts/UserContext";
