+++ app/(protected)/index.tsx
@@ -3,7 +3,7 @@ import { ScrollView, View } from "react-native";
 import { ActionButton } from "@/components/actions/ActionButton";
 import { ThemedText } from "@/components/text/ThemedText";
 import { TopView } from "@/feature/dashboard/components";
-import { useLinkInputModal } from "@/feature/links/contexts/LinkInputModalContext";
+import { useLinkInputModal } from "@/feature/links/application/contexts/LinkInputModalContext";
 import { LinksTopView } from "@/feature/links/views/LinksTopView";
 
 const FloatingButton = () => {
